# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: integer }
        companyId: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        budget: { type: ["number", "null"] }
        progress: { type: ["integer", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, companyId, name, status, priority, createdAt, updatedAt]
    ProjectInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        status: { type: string }
        priority: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        budget: { type: number }
        progress: { type: integer }
      required: [name]
    ProjectUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        status: { type: string }
        priority: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        budget: { type: number }
        progress: { type: integer }
    Task:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        projectId: { type: integer }
        projectName: { type: ["string", "null"] }
        assigneeId: { type: ["integer", "null"] }
        assigneeName: { type: ["string", "null"] }
        dueDate: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, title, status, priority, projectId, createdAt, updatedAt]
    TaskInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        status: { type: string }
        priority: { type: string }
        projectId: { type: integer }
        assigneeId: { type: integer }
        dueDate: { type: string }
      required: [title, projectId]
    TaskUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        status: { type: string }
        priority: { type: string }
        projectId: { type: integer }
        assigneeId: { type: integer }
        dueDate: { type: string }
...